fix: address PR #113 review blockers and medium/low issues
BLOCKER 1: remove docEntries byte-read loop and documents field from
CodexExecutor stdin JSON — Codex SDK 0.118.0 has no document block API,
base64-encoding 32 MB per task was wasteful and violated the hint-only
amendment from issue #112. Add one-time startup info-log instead.

BLOCKER 2: add Stage 2 post-download size guard in DownloadDocumentAsync
after GetInfoAndDownloadFile — Telegram's FileSize is advisory; actual
bytes may exceed the limit. Extract PdfTooLargeMessage() so both stages
share the same user-facing warning string.

Medium: pin @openai/codex-sdk@0.118.0 in CodexExecutor constructor
comment; remove dead per-message documents check from codex-bridge.mjs.

Low: MaxDocumentBytes int → long in TelegramOptions; drop tautology tests
(DocumentPath_FollowsSameChatMessageIndexShape,
PersistAttachments_False_DocumentKillSwitchRespected); replace arithmetic-
only size-cap test with two behaviour tests exercising the config value.

All 148 tests pass.

+++ src/Fleet.Agent/Configuration/AgentConfig.cs
@@ -63,7 +63,7 @@ public sealed class TelegramOptions
     public int AttachmentRetentionHours { get; set; } = 48;
 
     /// <summary>Maximum PDF document size in bytes; documents above this limit are rejected with a user-facing warning and not passed to the LLM. Default: 32 MB (Claude SDK per-document limit).</summary>
-    public int MaxDocumentBytes { get; set; } = 33_554_432; // 32 MB
+    public long MaxDocumentBytes { get; set; } = 33_554_432; // 32 MB
 }
 
 public sealed class RabbitMqOptions
